# EES-3976 - migrate to PNPM