--- a/azure-pipelines.dfe.yml
+++ b/azure-pipelines.dfe.yml
@@ -1,6 +1,9 @@
 variables:
   BuildConfiguration: 'Release'
   CI: true
+  PNPM_VERSION: 7.26.0
+  NODE_VERSION: 16.14.2
+  DOTNET_VERSION: 6.0.x
 
 trigger:
   branches:
@@ -16,12 +19,14 @@ jobs:
   - job: 'Backend'
     pool:
       vmImage: 'windows-2022'
+
     steps:
       - task: UseDotNet@2
         displayName: 'Install .NET Core SDK'
         inputs:
-          version: 6.0.x
+          version: $(DOTNET_VERSION)
           performMultiLevelLookup: true
+
       - task: 'DotNetCoreCLI@2'
         displayName: 'Restore'
         inputs:
@@ -29,11 +34,13 @@ jobs:
           projects: '**/GovUk.*/*csproj'
           feedsToUse: config
           nugetConfigPath: src/nuget.config
+
       - task: 'DotNetCoreCLI@2'
         displayName: 'Build'
         inputs:
           projects: '**/GovUk.*/*csproj'
           arguments: '--configuration $(BuildConfiguration)'
+
       - task: 'DotNetCoreCLI@2'
         displayName: 'Test'
         inputs:
@@ -49,6 +56,7 @@ jobs:
           projects: '**/GovUk.Education.ExploreEducationStatistics.Data.Api.csproj'
           arguments: '--self-contained true -r win-x64 --configuration $(BuildConfiguration) --output $(build.artifactstagingdirectory)/public-api'
           zipAfterPublish: True
+
       - task: 'DotNetCoreCLI@2'
         displayName: 'Publish Content API'
         inputs:
@@ -72,6 +80,7 @@ jobs:
           projects: '**/GovUk.Education.ExploreEducationStatistics.Notifier.csproj'
           arguments: '--configuration $(BuildConfiguration) --output $(build.artifactstagingdirectory)/notifier'
           zipAfterPublish: True
+
       - task: 'PublishPipelineArtifact@0'
         displayName: 'Publish Notifier Artifacts'
         inputs:
@@ -86,6 +95,7 @@ jobs:
           projects: '**/GovUk.Education.ExploreEducationStatistics.Publisher.csproj'
           arguments: '--configuration $(BuildConfiguration) --output $(build.artifactstagingdirectory)/publisher'
           zipAfterPublish: True
+
       - task: 'PublishPipelineArtifact@0'
         displayName: 'Publish Publisher Artifacts'
         inputs:
@@ -100,6 +110,7 @@ jobs:
           projects: '**/GovUk.Education.ExploreEducationStatistics.Data.Processor.csproj'
           arguments: '--configuration $(BuildConfiguration) --output $(build.artifactstagingdirectory)/processor'
           zipAfterPublish: True
+
       - task: 'PublishPipelineArtifact@0'
         displayName: 'Publish Processor Artifacts'
         inputs:
@@ -111,16 +122,26 @@ jobs:
       vmImage: 'ubuntu-20.04'
       demands:
         - azureps
+
     steps:
       - task: NodeTool@0
-        displayName: 'Install Node.js 16.14.2'
+        displayName: 'Install Node.js $(NODE_VERSION)'
         inputs:
-          versionSpec: '16.14.2'
+          versionSpec: $(NODE_VERSION)
+
+      - task: 'Bash@3'
+        displayName: 'Install PNPM'
+        inputs:
+          workingDir: '.'
+          targetType: 'inline'
+          script: 'npm i -g pnpm@$(PNPM_VERSION)'
+
       - task: UseDotNet@2
         displayName: 'Install .NET Core SDK'
         inputs:
-          version: 6.0.x
+          version: $(DOTNET_VERSION)
           performMultiLevelLookup: true
+
       - task: 'DotNetCoreCLI@2'
         displayName: 'Restore'
         inputs:
@@ -128,11 +149,13 @@ jobs:
           projects: '**/GovUk.Education.ExploreEducationStatistics.Admin.csproj'
           feedsToUse: config
           nugetConfigPath: src/nuget.config
+
       - task: 'DotNetCoreCLI@2'
         displayName: 'Build'
         inputs:
           projects: '**/GovUk.Education.ExploreEducationStatistics.Admin.csproj'
           arguments: '--configuration $(BuildConfiguration)'
+
       - task: 'DotNetCoreCLI@2'
         displayName: 'Test'
         inputs:
@@ -140,31 +163,23 @@ jobs:
           projects: '**/GovUk.Education.ExploreEducationStatistics.Admin.Tests.csproj'
           arguments: '--configuration $(BuildConfiguration) --collect "Code coverage"'
 
-      - task: Npm@1
-        displayName: 'npm run ci'
+      - task: 'Bash@3'
+        displayName: 'pnpm i'
         inputs:
-          command: custom
-          verbose: false
-          customCommand: ci
-      - task: Npm@1
-        displayName: 'npm run bootstrap'
+          targetType: 'inline'
+          script: 'pnpm i'
+
+      - task: 'Bash@3'
+        displayName: 'pnpm format'
         inputs:
-          command: custom
-          verbose: false
-          customCommand: 'run bootstrap'
-      - task: Npm@1
-        displayName: 'npm run format'
+          targetType: 'inline'
+          script: 'pnpm format'
+
+      - task: 'Bash@3'
+        displayName: 'pnpm run build'
         inputs:
-          command: custom
-          verbose: false
-          customCommand: 'run format'
-      - task: Npm@1
-        displayName: 'npm run build'
-        inputs:
-          command: custom
-          workingDir: 'src/explore-education-statistics-admin'
-          verbose: false
-          customCommand: 'run build'
+          targetType: 'inline'
+          script: 'pnpm --filter=explore-education-statistics-admin run build'
 
       - task: CopyFiles@2
         displayName: 'Copy files to wwwroot'
@@ -179,6 +194,7 @@ jobs:
           publishWebProjects: false
           projects: '**/GovUk.Education.ExploreEducationStatistics.Admin.csproj'
           arguments: '--self-contained true -r win-x64 --configuration $(BuildConfiguration) --output $(build.artifactstagingdirectory)'
+
       - task: 'PublishPipelineArtifact@0'
         displayName: 'Publish Admin Artifact'
         inputs:
@@ -189,65 +205,55 @@ jobs:
     pool:
       vmImage: 'ubuntu-20.04'
       demands: azureps
+
     steps:
       - task: NodeTool@0
-        displayName: 'Install Node.js 16.14.2'
+        displayName: 'Install Node.js $(NODE_VERSION)'
         inputs:
-          versionSpec: '16.14.2'
-      - task: 'Npm@1'
-        displayName: 'npm run ci'
+          versionSpec: $(NODE_VERSION)
+
+      - task: 'Bash@3'
+        displayName: 'Install PNPM'
         inputs:
-          command: 'custom'
           workingDir: '.'
-          verbose: false
-          customCommand: 'ci'
-      - task: 'Npm@1'
-        displayName: 'npm run bootstrap'
+          targetType: 'inline'
+          script: 'npm i -g pnpm@$(PNPM_VERSION)'
+
+      - task: 'Bash@3'
+        displayName: 'pnpm i'
         inputs:
-          command: 'custom'
           workingDir: '.'
-          verbose: false
-          customCommand: 'run bootstrap'
-      - script: 'npm audit --json >> audit.json'
-        workingDirectory: 'src/explore-education-statistics-frontend'
-        displayName: 'Run npm audit'
-        continueOnError: true
-        enabled: false
-      - task: PublishPipelineArtifact@0
-        displayName: 'Publish Pipeline Artifact audit_results'
+          targetType: 'inline'
+          script: 'pnpm i'
+
+      - task: 'Bash@3'
+        displayName: 'pnpm tsc'
         inputs:
-          artifactName: 'audit_results'
-          targetPath: 'src/explore-education-statistics-frontend/audit.json'
-        continueOnError: true
-        enabled: false
-      - task: 'Npm@1'
-        displayName: 'npm run tsc'
-        inputs:
-          command: 'custom'
           workingDir: '.'
-          verbose: false
-          customCommand: 'run tsc'
-      - task: 'Npm@1'
-        displayName: 'npm run lint'
+          targetType: 'inline'
+          script: 'pnpm tsc'
+
+      - task: 'Bash@3'
+        displayName: 'pnpm lint'
         inputs:
-          command: 'custom'
           workingDir: '.'
-          verbose: false
-          customCommand: 'run lint'
-      - task: 'Npm@1'
-        displayName: 'npm run format:check'
+          targetType: 'inline'
+          script: 'pnpm lint'
+
+      - task: 'Bash@3'
+        displayName: 'pnpm format:check'
         inputs:
-          command: 'custom'
           workingDir: '.'
-          verbose: false
-          customCommand: 'run format:check'
-      - task: 'Npm@1'
-        displayName: 'npm run test:ci'
+          targetType: 'inline'
+          script: 'pnpm format:check'
+
+      - task: 'Bash@3'
+        displayName: 'pnpm test:ci'
         inputs:
-          command: 'custom'
           workingDir: '.'
-          verbose: false
-          customCommand: 'run test:ci'
+          targetType: 'inline'
+          script: 'pnpm test:ci'
+
       - task: 'PublishTestResults@2'
         displayName: 'Publish frontend test results'
         inputs:
@@ -256,50 +262,34 @@ jobs:
           searchFolder: './src'
           testRunTitle: 'Release Jest tests'
           mergeTestResults: true
-      - task: 'Npm@1'
-        displayName: 'Build frontend'
+
+      - task: 'Bash@3'
+        displayName: 'pnpm run build'
         inputs:
-          command: 'custom'
-          workingDir: './src/explore-education-statistics-frontend'
-          verbose: false
-          customCommand: 'run build'
+          targetType: 'inline'
+          script: 'pnpm --filter=explore-education-statistics-frontend run build'
+
       - task: 'Bash@3'
         displayName: 'Remove dev dependencies'
         inputs:
           targetType: 'inline'
           script: 'rm -rf ./src/**/node_modules'
-      - task: 'Npm@1'
+
+      - task: 'Bash@3'
         displayName: 'Install prod dependencies'
         inputs:
-          command: 'custom'
-          workingDir: '.'
-          verbose: false
-          customCommand: 'run bootstrap:prod'
+          targetType: 'inline'
+          script: 'pnpm i --filter=explore-education-statistics-frontend --prod'
+
       - task: 'ArchiveFiles@2'
         displayName: 'Archive frontend'
         inputs:
           rootFolderOrFile: './src/explore-education-statistics-frontend'
           includeRootFolder: false
           archiveFile: '$(Build.ArtifactStagingDirectory)/public-frontend.zip'
+
       - task: 'PublishPipelineArtifact@0'
         displayName: 'Publish Pipeline Artifact'
         inputs:
           artifactName: 'public-frontend'
           targetPath: '$(Build.ArtifactStagingDirectory)'
-
-  - job: 'MiscellaneousArtifacts'
-    pool:
-      vmImage: 'ubuntu-20.04'
-    steps:
-      - task: CopyFiles@2
-        displayName: 'Copy Pipfiles to tests'
-        inputs:
-          Contents: |
-            Pipfile
-            Pipfile.lock
-          TargetFolder: 'tests'
-      - task: PublishPipelineArtifact@0
-        displayName: 'Publish test files'
-        inputs:
-          artifactName: 'tests'
-          targetPath: 'tests'