--- a/src/explore-education-statistics-common/src/modules/charts/components/__tests__/MapBlockInternal.test.tsx
+++ b/src/explore-education-statistics-common/src/modules/charts/components/__tests__/MapBlockInternal.test.tsx
@@ -51,7 +51,7 @@ describe('MapBlockInternal', () => {
 
     await waitFor(() => {
       const paths = container.querySelectorAll<HTMLElement>(
-        '.leaflet-container svg path',
+        '.leaflet-container svg:not(.leaflet-attribution-flag) path',
       );
 
       expect(paths).toHaveLength(4);
@@ -172,9 +172,8 @@ describe('MapBlockInternal', () => {
       );
 
       userEvent.selectOptions(select, select.children[1] as HTMLElement);
-
       const paths = container.querySelectorAll<HTMLElement>(
-        '.leaflet-container svg path',
+        '.leaflet-container svg:not(.leaflet-attribution-flag) path',
       );
 
       expect(paths).toHaveLength(4);
@@ -221,9 +220,8 @@ describe('MapBlockInternal', () => {
     expect(group1Options[0]).toHaveTextContent('Leeds');
 
     userEvent.selectOptions(select, group1Options[0]);
-
     const paths = container.querySelectorAll<HTMLElement>(
-      '.leaflet-container svg path',
+      '.leaflet-container svg:not(.leaflet-attribution-flag) path',
     );
 
     expect(paths).toHaveLength(4);
@@ -342,7 +340,7 @@ describe('MapBlockInternal', () => {
     userEvent.selectOptions(select, group1Options[0]);
 
     const paths = container.querySelectorAll<HTMLElement>(
-      '.leaflet-container svg path',
+      '.leaflet-container svg:not(.leaflet-attribution-flag) path',
     );
 
     expect(paths[3]).toHaveClass('selected');
@@ -380,7 +378,7 @@ describe('MapBlockInternal', () => {
     });
 
     const paths = container.querySelectorAll<HTMLElement>(
-      '.leaflet-container svg path',
+      '.leaflet-container svg:not(.leaflet-attribution-flag) path',
     );
 
     expect(paths).toHaveLength(4);